+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+* Sorting for contacts
+
 ## [0.1.0] - 2025-05-23
 
 ### Added